+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Next version
 + Generate token from token
 + Add Refreh Token route
++ Get expiry date from token
 
 
 ## 0.11.0